% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyMatch.R
\name{familyMatch}
\alias{familyMatch}
\title{Split the names with or without author names into the required format}
\usage{
familyMatch(
  genuslist = NULL,
  splist = NULL,
  taxon = c("plant", "fern", "bird", "mammal", "amphibian", "fish")
)
}
\arguments{
\item{genuslist}{A character vector of the input genus names.}

\item{splist}{A character vector of the input speices names.}

\item{taxon}{Select one specific taxon group to match. Currently, 'plant','fern','bird','mammal','amphibian','fish'.}
}
\value{
A data frame with three or four columns: "sorter", "sp", "genus" and "family".
}
\description{
This function is used to format the input taxon names before running the function \code{\link[U.Taxonstand:nameMatch]{nameMatch}}.
}
\examples{
# genus names
familyMatch(genuslist=c("Aa","Pinus","Humbertodendron"), taxon="plant")
familyMatch(genuslist = c("Acanthocharax","Agosia","Alticorpus"),taxon="fish")

# species names
familyMatch(splist=c("Aa xyz","Pinus xyz","Humbertodendron xyz"), taxon="plant")
familyMatch(splist = c("Acanthocharax xyz","Agosia xyz","Alticorpus xyz"),taxon="fish")

}
\author{
Jian Zhang
}
