% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genusFamily_Amphibians.R
\docType{data}
\name{genusFamily_Amphibians}
\alias{genusFamily_Amphibians}
\title{A data of the genus and family list for plants.}
\format{
An object of class \code{data.frame} with 1234 rows and 2 columns.
}
\usage{
data(genusFamily_Amphibians)
}
\description{
A data frame with 2 variables.
}
\keyword{datasets}
