% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genusFamily_Birds.R
\docType{data}
\name{genusFamily_Birds}
\alias{genusFamily_Birds}
\title{A data of the genus and family list for plants.}
\format{
An object of class \code{data.frame} with 2875 rows and 2 columns.
}
\usage{
data(genusFamily_Birds)
}
\description{
A data frame with 2 variables.
}
\keyword{datasets}
