% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genusFamily_Ferns.R
\docType{data}
\name{genusFamily_Ferns}
\alias{genusFamily_Ferns}
\title{A data of the genus and family list for plants.}
\format{
An object of class \code{data.frame} with 1014 rows and 2 columns.
}
\usage{
data(genusFamily_Ferns)
}
\description{
A data frame with 2 variables.
}
\keyword{datasets}
