% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genusFamily_Fishes.R
\docType{data}
\name{genusFamily_Fishes}
\alias{genusFamily_Fishes}
\title{A data of the genus and family list for plants.}
\format{
An object of class \code{data.frame} with 9794 rows and 2 columns.
}
\usage{
data(genusFamily_Fishes)
}
\description{
A data frame with 2 variables.
}
\keyword{datasets}
