% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genusFamily_Mammals.R
\docType{data}
\name{genusFamily_Mammals}
\alias{genusFamily_Mammals}
\title{A data of the genus and family list for plants.}
\format{
An object of class \code{data.frame} with 1805 rows and 2 columns.
}
\usage{
data(genusFamily_Mammals)
}
\description{
A data frame with 2 variables.
}
\keyword{datasets}
