% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genusFamily_Plants.R
\docType{data}
\name{genusFamily_Plants}
\alias{genusFamily_Plants}
\title{A data of the genus and family list for plants.}
\format{
An object of class \code{data.frame} with 31262 rows and 2 columns.
}
\usage{
data(genusFamily_Plants)
}
\description{
A data frame with 2 variables.
}
\keyword{datasets}
