% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista.r
\docType{data}
\name{lista}
\alias{lista}
\title{A data file of fuzzy genus names of vascular plants}
\format{
An object of class \code{data.frame} with 3804 rows and 52 columns.
}
\usage{
data(lista)
}
\description{
The data file is from the R package Taxonstand
}
\keyword{datasets}
