% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameMatch.r
\name{nameMatch}
\alias{nameMatch}
\title{The main function for standardizing a list of input taxon}
\usage{
nameMatch(
  spList = NULL,
  spSource = NULL,
  author = TRUE,
  max.distance = 1,
  isPlant = TRUE,
  genusPairs = NULL
)
}
\arguments{
\item{spList}{A data frame or a character vector specifying the input taxon. An example of data frame could be found in the example data 'spExample'.A character vector specifying the input taxa, each element including genus and specific epithet and, potentially, author name and infraspecific abbreviation and epithet.}

\item{spSource}{A data frame of the backbone database with standarized taxon names. An example of data frame could be found in the example data 'databaseExample'. Multiple backbone databases for plants and animals that can be directly used by U.Taxonstand are available online (https://github.com/nameMatch/Data_backbone). They include global taxonomic databases for vascular plants, amphibians, birds, fishes, mammals, and reptiles. Global backbone databases for other groups of organisms may be placed on the website in future.}

\item{author}{Logical. If TRUE (default), the function tries to match author names from the input taxon and calculate the author distance.}

\item{max.distance}{A number indicating the maximum distance allowed for a match in \code{\link[base:agrep]{agrep}} when performing corrections of spelling errors in specific epithets.}

\item{isPlant}{Logical. If TRUE(default), the function tried to use one datafile from R package Taxonstand for fuzzy matching of genus names.}
}
\value{
A data frame with the following columns: \itemize{
   \item{\emph{Submitted_Name}}{: Original taxon name as provided in input.}
   \item{\emph{Submitted_Author}}{: Original author name as provided in input.}
   \item{\emph{Submitted_Genus}}{: Original genus name as provided in input.}
   \item{\emph{Submitted_Rank}}{: Original taxonomic rank as provied in input.}
   \item{\emph{Name_in_database}}{: Matched latin name, extracted from the spSource.}
   \item{\emph{Author_in_database}}{: Matched author name, extracted from the spSource.}
   \item{\emph{Genus_in_database}}{: Matched genus name, extracted from the spSource.}
   \item{\emph{Rank_in_database}}{: Matched taxonomic rank of specific epithet.}
   \item{\emph{ID_in_database}}{: Matched ID, extracted from the spSource.}
   \item{\emph{Name_set}}{: Gives a rank based on the column Score. If you select the rows with Name_set = 1, only one of the 'best' matched names for each input taxon includes.}
   \item{\emph{Fuzzy}}{: ‘TRUE’ means the matching result is fully or partially based on fuzzy matching algorithm; ‘FALSE’ means the matching result is not based on fuzzy matching algorithm.}
   \item{\emph{Score}}{: Matching score (ranging from 1 to 0, when author of a scientific name is provided in user’s species list, or ranging from 0 to 0.8 when author of a scientific name is not provided). A larger value represents a better matching.}
   \item{\emph{name.dist}}{: The distance between Submitted_Name and Name_in_database.}
   \item{\emph{author.dist}}{: The distance between Submitted_Author and Author_in_database.}
   \item{\emph{New_name}}{: The accepted name in the spSource.In the case that this column is empty and the column Score has a value > 0, the matched name is treated as an accepted name in backbone database. In the case that the matched name is treated as a synonym in backbone database, the name shown in this column is the accepted name of the synonym based on backbone database. For a name treated as an unresolved name or as a synonym whose accepted name cannot be identified on backbone database, the text “[Accepted name needs to be determined]” is shown in this column.}
   \item{\emph{New_author}}{: Authority of the name in the column New_name based on backbone database.}
   \item{\emph{New_ID}}{: ID of the name in the column New_name based on backbone database.}
   \item{\emph{Family}}{: Family name, extracted from the spSource.}
   \item{\emph{Name_spLev}}{: If the input taxon name cannot be matched at the same rank level, the species-level matching name shows here.}
   \item{\emph{Accepted_SPNAME}}{: If the name in the column “Submitted_Name” is matched with a name in the backbone database, either as an accepted name or a synonym, the species name of the accepted name of the matched name will be listed here.}
   \item{\emph{NOTE}}{: This column contains some notes, e.g. no matching or multiple matching results for a given name, or only matching at species level for a trinomial name.}
}
}
\description{
Replacing synonyms by accepted names and removing orthographical errors in the raw names
}
\examples{
## load the database
data(databaseExample)

## The input names as a character vector
sps <- c("Syntoma comosum (L.) Dalla Torre & Sarnth.", "Turczaninowia fastigiata (Fisch.) DC.","Zizyphora abd-el-asisii Hand.-Mazz.")
nameMatch(spList=sps, spSource=databaseExample, author = TRUE, max.distance= 1)

## The input names as a data frame
data(spExample)
res <- nameMatch(spList=spExample, spSource=databaseExample, author = TRUE, max.distance= 1)
head(res)

## Using the additional data of genus pairs for fuzzy matching of genus names
data(spExample)
data(genusPairs_Plants)
res <- nameMatch(spList=spExample, spSource=databaseExample, author = TRUE, max.distance= 1, isPlant=TRUE, genusPairs=genusPairs_Plants)
head(res)

}
\references{
Zhang, J. & Qian, H. U.Taxonstand: An R package for standardizing scientific names of plants and animals. In prep.
}
\author{
Jian Zhang & Hong Qian
}
