% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameMatch_ck.r
\name{nameMatch_ck}
\alias{nameMatch_ck}
\title{The subfunction for standardizing one input taxon}
\usage{
nameMatch_ck(spData, spSource = NULL, author = TRUE, max.distance = 1)
}
\arguments{
\item{spData}{A character vector specifying the input taxon. The binominal name and the author name are connected by "#", e.g., "Hieracium aragonense#Scheele" and "Pilosella X subulatissima#(Zahn) Mateo".}

\item{spSource}{A data frame of the database with standarized taxon names. An example of data frame could be found in the example data 'databaseExample'.}

\item{author}{Logical. If TRUE (default), the function tries to match author names from the input taxon and calculate the author distance.}

\item{max.distance}{A number indicating the maximum distance allowed for a match in \code{\link[base:agrep]{agrep}} when performing corrections of spelling errors in specific epithets.}
}
\value{
A data frame with the following columns: \itemize{
   \item{\emph{Submitted_Name_Author}}{: The combination of the original taxon name and author name as provided in input.}
   \item{\emph{Name_in_database}}{: Matched specific epithet, extracted from the spSource.}
   \item{\emph{Author_in_database}}{: Matched author name, extracted from the spSource.}
   \item{\emph{Genus_in_database}}{: Matched genus name, extracted from the spSource.}
   \item{\emph{Rank_in_database}}{: Matched taxonomic rank of specific epithet.}
   \item{\emph{ID_in_database}}{: Matched ID, extracted from the spSource.}
   \item{\emph{NOTE}}{: This column contains some notes, e.g. no matching or multiple matching results for a given name, or only matching at species level for a trinomial name.}
   \item{\emph{Name_spLev}}{: If the input taxon name cannot be matched at the same rank level, the species-level matching name shows here.}
   \item{\emph{ACCEPTED_ID}}{: The accepted ID in the spSource.}
   \item{\emph{FAMILY}}{: Family name, extracted from the spSource.}
}
}
\description{
Replacing synonyms by accepted names and removing orthographical errors. This function is used to run the main function "nameMatch", we suggest NOT to use it independently.
}
\author{
Jian Zhang & Hong Qian
}
