% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameMatch_WCVP.R
\name{nameMatch_WCVP}
\alias{nameMatch_WCVP}
\title{The function for standardizing a list of input plant names using the World Checklist of Vascular Plants (WCVP)}
\usage{
nameMatch_WCVP(
  spList = NULL,
  author = TRUE,
  max.distance = 1,
  genusPairs = NULL,
  Append = FALSE
)
}
\arguments{
\item{spList}{A data frame or a character vector specifying the input taxon. An example of data frame could be found in the example data 'spExample'.A character vector specifying the input taxa, each element including genus and specific epithet and, potentially, author name and infraspecific abbreviation and epithet.}

\item{author}{Logical. If TRUE (default), the function tries to match author names from the input taxon and calculate the author distance.}

\item{max.distance}{A number indicating the maximum distance allows for a match in \code{\link[base:agrep]{agrep}} when performing corrections of spelling errors in specific epithets.}

\item{genusPairs}{Some genera have one or more variants in spelling (e.g. Euodia versus Evodia, Eccremis versus Excremis, Ziziphus versus Zizyphus). When a list of such genera is available, U.Taxonstand can use the information in the list as additional data to match names between the user’s species list and the taxonomic database. The data file should a dataframe with two columns (Genus01 and Genus02).}

\item{Append}{Logical. If TRUE, the function will add other columns (e.g., geographic distribution and common names in different language) in the database into the final result.}
}
\value{
A data frame with the following columns: \itemize{
   \item{\emph{Submitted_Name}}{: Original taxon name as provided in input.}
   \item{\emph{Submitted_Author}}{: Original author name as provided in input.}
   \item{\emph{Submitted_Genus}}{: Original genus name as provided in input.}
   \item{\emph{Submitted_Rank}}{: Original taxonomic rank as provied in input.}
   \item{\emph{Name_in_database}}{: Matched latin name, extracted from the spSource.}
   \item{\emph{Author_in_database}}{: Matched author name, extracted from the spSource.}
   \item{\emph{Genus_in_database}}{: Matched genus name, extracted from the spSource.}
   \item{\emph{Rank_in_database}}{: Matched taxonomic rank of specific epithet.}
   \item{\emph{ID_in_database}}{: Matched ID, extracted from the spSource.}
   \item{\emph{Name_set}}{: Gives a rank based on the column Score. If you select the rows with Name_set = 1, only one of the 'best' matched names for each input taxon includes.}
   \item{\emph{Fuzzy}}{: ‘TRUE’ means the matching result is fully or partially based on fuzzy matching algorithm; ‘FALSE’ means the matching result is not based on fuzzy matching algorithm.}
   \item{\emph{Score}}{: Matching score (ranging from 1 to 0, when author of a scientific name is provided in user’s species list, or ranging from 0 to 0.8 when author of a scientific name is not provided). A larger value represents a better matching.}
   \item{\emph{name.dist}}{: The distance between Submitted_Name and Name_in_database.}
   \item{\emph{author.dist}}{: The distance between Submitted_Author and Author_in_database.}
   \item{\emph{New_name}}{: The accepted name in the spSource.In the case that this column is empty and the column Score has a value > 0, the matched name is treated as an accepted name in the database. In the case that the matched name is treated as a synonym in the database, the name shown in this column is the accepted name of the synonym based on the database. For a name treated as an unresolved name or as a synonym whose accepted name cannot be identified on the database, the text “[Accepted name needs to be determined]” is shown in this column.}
   \item{\emph{New_author}}{: Authority of the name in the column New_name based on the database.}
   \item{\emph{New_ID}}{: ID of the name in the column New_name based on the database.}
   \item{\emph{Family}}{: Family name, extracted from the spSource.}
   \item{\emph{Name_spLev}}{: If the input taxon name cannot be matched at the same rank level, the species-level matching name shows here.}
   \item{\emph{Accepted_SPNAME}}{: If the name in the column “Submitted_Name” is matched with a name in the database, either as an accepted name or a synonym, the species name of the accepted name of the matched name will be listed here.}
   \item{\emph{NOTE}}{: This column contains some notes, e.g. no matching or multiple matching results for a given name, or only matching at species level for a trinomial name.}
}
}
\description{
Replacing synonyms by accepted names and removing orthographical errors in the raw names. This function is used to match with the World Checklist of Vascular Plants (WCVP). The function was developed based on the database of the R package rWCVPdata.
}
\examples{

## The input names as a character vector
nameMatch_WCVP(spList="Cyclobalanopsis myrsinifolia (Blume) Oerst.")
nameMatch_WCVP(spList=c("Cyclobalanopsis myrsinifolia (Blume) Oerst.",
"Cyclobalanopsis myrsinifolium",
"Cyclobalanopsis mrrsinifolia",
"Cyclobalamopsis mrrsinifolia"))

## Using the additional data of genus pairs for fuzzy matching of genus names
data(genusPairs_Plants)
nameMatch_WCVP(spList=c("Cyclobalanopsis myrsinifolia (Blume) Oerst.",
"Cyclobalanopsis myrsinifolium",
"Cyclobalanopsis mrrsinifolia",
"Cyclobalamopsis myrsinifolia",
"Evodia chaffanjonii","Evodia lyi","Euodia lyi"),
genusPairs=genusPairs_Plants)


}
\references{
Govaerts, R., Nic Lughadha, E. et al. The World Checklist of Vascular Plants, a continuously updated resource for exploring global plant diversity. Sci Data 8, 215 (2021). https://doi.org/10.1038/s41597-021-00997-6
}
