% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameSplit.r
\name{nameSplit}
\alias{nameSplit}
\title{Split the names with or without author names into the required format}
\usage{
nameSplit(splist)
}
\arguments{
\item{splist}{A character vector of the input taxon, e.g., "Hieracium aragonense Scheele" and "Pilosella X subulatissima (Zahn) Mateo".}
}
\value{
A data frame with three columns: "Name", "Author" and "Rank".
}
\description{
This function is used to format the input taxon names before running the function \code{\link[U.Taxonstand:nameMatch]{nameMatch}}.
}
\examples{
sps <- c("Syntoma comosum (L.) Dalla Torre & Sarnth.", "Eupatorium betoniciforme f. alternifolium Hicken", "Turczaninowia fastigiata (Fisch.) DC.","Zizyphora abd-el-asisii Hand.-Mazz.","Baccharis X paulopolitana I.L.Teodoro & W.Hoehne","Accipiter albogularis woodfordi (Sharpe, 1888)")
(splist <- nameSplit(splist=sps))

}
\author{
Jian Zhang & Hong Qian
}
