% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spExample.r
\docType{data}
\name{spExample}
\alias{spExample}
\title{An example data of the input name list}
\format{
A data frame with 3 variables: \itemize{
   \item{\emph{Name}}{: Original taxon name as provided in input.}
   \item{\emph{Author}}{: Original author name as provided in input.}
   \item{\emph{Rank}}{: Optional. Original taxonomic rank as provied in input.}
}
}
\usage{
spExample
}
\description{
A data frame with 289 rows and 3 variables.
}
\examples{
data(spExample)
head(spExample)

}
\keyword{datasets}
