% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synSearch.R
\name{synSearch}
\alias{synSearch}
\title{List all the synonyms and accepted names for your input names}
\usage{
synSearch(splist, spSource)
}
\arguments{
\item{splist}{A character vector of the input taxon, e.g., "Hieracium aragonense Scheele" and "Pilosella X subulatissima (Zahn) Mateo".}

\item{spSource}{A data frame of the database with standarized taxon names. An example of data frame could be found in the example data 'databaseExample'. Multiple databases for plants and animals that can be directly used by U.Taxonstand are available online (https://github.com/nameMatch/Database). They include global taxonomic databases for bryophytes, vascular plants, amphibians, birds, fishes, mammals, reptile and others. You can creat your own database for name matching.}
}
\value{
A data frame with the columns of all synonyms and accepted names.
}
\description{
NOTE that this function only works for the matched taxon names. Otherwise, you should use the function \code{\link[U.Taxonstand:nameMatch]{nameMatch}} to do the matching first.
}
\examples{
data(databaseExample)
synSearch(splist="Isoetes tenella", spSource=databaseExample)
synSearch(splist=c("Isoetes tenella","Isoetes echinospora","Isoetes underwoodii L.F. Hend. Isoetes"), spSource=databaseExample)

}
\author{
Jian Zhang
}
