% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameMatch.r
\name{nameMatch}
\alias{nameMatch}
\title{The main function for standardizing a list of input taxon}
\usage{
nameMatch(
  spList = NULL,
  spSource = NULL,
  author = TRUE,
  max.distance = 1,
  genusPairs = NULL,
  matchFirst = TRUE,
  Append = FALSE
)
}
\arguments{
\item{spList}{A data frame or a character vector specifying the input taxon. An example of data frame could be found in the example data 'spExample'.A character vector specifying the input taxa, each element including genus and specific epithet and, potentially, author name and infraspecific abbreviation and epithet.}

\item{spSource}{A data frame of the database with standarized taxon names. An example of data frame could be found in the example data 'databaseExample'. Multiple databases for plants and animals that can be directly used by U.Taxonstand are available online (https://github.com/nameMatch/Database). They include global taxonomic databases for bryophytes, vascular plants, amphibians, birds, fishes, mammals, and reptiles. Global databases for other taxon groups of organisms may be placed on the website in future.}

\item{author}{Logical. If TRUE (default), the function tries to match author names from the input taxon and calculate the author distance.}

\item{max.distance}{A number indicating the maximum distance allows for a match in \code{\link[base:agrep]{agrep}} when performing corrections of spelling errors in specific epithets.}

\item{genusPairs}{Some genera have one or more variants in spelling (e.g. Euodia versus Evodia, Eccremis versus Excremis, Ziziphus versus Zizyphus). When a list of such genera is available, U.Taxonstand can use the information in the list as additional data to match names between the user’s species list and the taxonomic database. The data file should a dataframe with two columns (Genus01 and Genus02).}

\item{matchFirst}{Logical. If TRUE, the function only keeps the first 'BEST' matching result for each input taxon. If FALST, all matching results are listed.}

\item{Append}{Logical. If TRUE, the function will add other columns (e.g., geographic distribution and common names in different language) in the database into the final result.}
}
\value{
A data frame with the following columns: \itemize{
   \item{\emph{Submitted_Name}}{: Original taxon name as provided in input.}
   \item{\emph{Submitted_Author}}{: Original author name as provided in input.}
   \item{\emph{Submitted_Genus}}{: Original genus name as provided in input.}
   \item{\emph{Submitted_Rank}}{: Original taxonomic rank as provied in input.}
   \item{\emph{Name_in_database}}{: Matched latin name, extracted from the spSource.}
   \item{\emph{Author_in_database}}{: Matched author name, extracted from the spSource.}
   \item{\emph{Genus_in_database}}{: Matched genus name, extracted from the spSource.}
   \item{\emph{Rank_in_database}}{: Matched taxonomic rank of specific epithet.}
   \item{\emph{ID_in_database}}{: Matched ID, extracted from the spSource.}
   \item{\emph{Name_set}}{: Gives a rank based on the column Score. If you select the rows with Name_set = 1, only one of the 'best' matched names for each input taxon includes.}
   \item{\emph{Fuzzy}}{: ‘TRUE’ means the matching result is fully or partially based on fuzzy matching algorithm; ‘FALSE’ means the matching result is not based on fuzzy matching algorithm.}
   \item{\emph{Score}}{: Matching score (ranging from 1 to 0, when author of a scientific name is provided in user’s species list, or ranging from 0 to 0.8 when author of a scientific name is not provided). A larger value represents a better matching.}
   \item{\emph{name.dist}}{: The distance between Submitted_Name and Name_in_database.}
   \item{\emph{author.dist}}{: The distance between Submitted_Author and Author_in_database.}
   \item{\emph{New_name}}{: The accepted name in the spSource.In the case that this column is empty and the column Score has a value > 0, the matched name is treated as an accepted name in the database. In the case that the matched name is treated as a synonym in the database, the name shown in this column is the accepted name of the synonym based on the database. For a name treated as an unresolved name or as a synonym whose accepted name cannot be identified on the database, the text “[Accepted name needs to be determined]” is shown in this column.}
   \item{\emph{New_author}}{: Authority of the name in the column New_name based on the database.}
   \item{\emph{New_ID}}{: ID of the name in the column New_name based on the database.}
   \item{\emph{Family}}{: Family name, extracted from the spSource.}
   \item{\emph{Name_spLev}}{: If the input taxon name cannot be matched at the same rank level, the species-level matching name shows here.}
   \item{\emph{Accepted_SPNAME}}{: If the name in the column “Submitted_Name” is matched with a name in the database, either as an accepted name or a synonym, the species name of the accepted name of the matched name will be listed here.}
   \item{\emph{NOTE}}{: This column contains some notes, e.g. no matching or multiple matching results for a given name, or only matching at species level for a trinomial name.}
}
}
\description{
Replacing synonyms by accepted names and removing orthographical errors in the raw names
}
\examples{
## load the database
data(databaseExample)

## The input names as a character vector
sps <- c("Syntoma comosum (L.) Dalla Torre & Sarnth.", "Turczaninowia fastigiata (Fisch.) DC.",
"Zizyphora abd-el-asisii Hand.-Mazz.")
nameMatch(spList=sps, spSource=databaseExample, author = TRUE, max.distance= 1, matchFirst=TRUE)

## The input names as a data frame
data(spExample)
res <- nameMatch(spList=spExample, spSource=databaseExample, author = TRUE, max.distance= 1,
matchFirst=TRUE)
head(res)

## Using the additional data of genus pairs for fuzzy matching of genus names
data(spExample)
data(genusPairs_Plants)
res <- nameMatch(spList=spExample, spSource=databaseExample, author = TRUE, max.distance= 1, 
genusPairs=genusPairs_Plants, matchFirst=TRUE)
head(res)

## Species name matching for bird species (Not run)
# load the formatted ITIS Aves database; the formatted database can be downloaded here:
# https://github.com/nameMatch/Database
# require(openxlsx)
# databaseAves <- read.xlsx("Birds_ITIS_database.xlsx")
# nameMatch(spList=c("Hieraaetus fasciata","Merops leschenaultia","Egretta sacra",
# "Sturnia philippensis","Phoenicurus caeruleocephala","Enicurus maculates",
# "Orthotomus cucullatus","Phalacrocorax carbo"),
# spSource=databaseAves, author = FALSE, max.distance= 1)

## Adding other information in the database into the final result (Not run)
# sps <- c("Syntoma comosum (L.) Dalla Torre & Sarnth.", "Turczaninowia fastigiata (Fisch.) DC.",
# "Zizyphora abd-el-asisii Hand.-Mazz.")
# nameMatch(spList=sps, spSource=China_Plants_SP2000_2022Edition, author = TRUE, max.distance= 1,
# genusPairs=genusPairs_Plants, Append=TRUE)
# NOTE: The raw data "China_Plants_SP2000_2022Edition" can be downloaded here:
# https://www.plantplus.cn/doi/10.12282/plantdata.0061;
# the formatted database "Plants_SP2000_2022Edition_China" can be downloaded here:
# https://github.com/nameMatch/Database

}
\references{
Zhang, J. & Qian, H. (2022). U.Taxonstand: An R package for standardizing scientific names of plants and animals. Plant Diversity, 45(1): 1-5. DOI: 10.1016/j.pld.2022.09.001.
}
\author{
Jian Zhang & Hong Qian
}
